+++ src/sources/crd.ts
@@ -1,8 +1,8 @@
 import { ExternaldnsV1alpha1Api, KubernetesClient } from '../deps.ts';
 
 import type { CrdSourceConfig } from "../common/config.ts";
-import type { DnsSource, SourceRecord, PlainRecordData } from "../common/contract.ts";
-import { WatchLister } from "../common/watch-lister.ts";
+import type { DnsSource, SourceRecord, PlainRecordData } from "../common/types.ts";
+import { WatchLister } from "./lib/watch-lister.ts";
 
 export class CrdSource implements DnsSource {
