+++ src/sources/node.ts
@@ -1,8 +1,8 @@
 import { CoreV1Api, KubernetesClient } from '../deps.ts';
 
 import type { NodeSourceConfig } from "../common/config.ts";
-import type { DnsSource, SourceRecord } from "../common/contract.ts";
-import { WatchLister } from "../common/watch-lister.ts";
+import type { DnsSource, SourceRecord } from "../common/types.ts";
+import { WatchLister } from "./lib/watch-lister.ts";
 
 import { splitIntoV4andV6 } from "../dns-logic/endpoints.ts";
